backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
twilio==8.4.1